Clang-tidy-check-full fix
Issue: #ICSTXN

Testing:'ninja all tests'

+++ static_core/runtime/entrypoints/entrypoints.cpp
@@ -1942,7 +1942,7 @@ extern "C" coretypes::String *CoreStringConcat4(coretypes::String *str1, coretyp
 
 extern "C" uint8_t CoreStringEquals(coretypes::String *str1, coretypes::String *str2)
 {
-    return coretypes::String::StringsAreEqual(str1, str2);
+    return static_cast<uint8_t>(coretypes::String::StringsAreEqual(str1, str2));
 }
 
 }  // namespace ark
